refactor: 출석 정책 클래스 분리
