+++ src/test/java/AttendanceRecordTest.java
@@ -65,22 +65,10 @@ public class AttendanceRecordTest {
     }
 
     @Test
-    void 출석_날짜가_주말이면_예외가_발생한다() {
-        assertThatThrownBy(() -> new AttendanceRecord(() -> weekend).attend("09:59"))
-                .isInstanceOf(IllegalArgumentException.class);
-    }
-
-    @Test
-    void 캠퍼스_운영_시간이_아니면_예외가_발생한다() {
-        assertThatThrownBy(() -> new AttendanceRecord(() -> weekday).attend("07:59"))
-                .isInstanceOf(IllegalArgumentException.class)
-                .hasMessage("[ERROR] 캠퍼스 운영 시간이 아닙니다.");
-    }
-
-    @Test
     void 이미_출석한_날짜이면_예외가_발생한다() {
         AttendanceRecord attendanceRecord = new AttendanceRecord(() -> weekday);
-        
+        attendanceRecord.attend("09:59");
+
         assertThatThrownBy(() -> attendanceRecord.attend("10:06"))
                 .isInstanceOf(IllegalArgumentException.class)
                 .hasMessage("[ERROR] 해당 날짜에는 이미 출석했습니다. 수정 기능을 이용해주세요.");
